requirements.txt
# Django Backend
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Telegram Bot
aiogram>=3.0.0
aiohttp>=3.9.0

